+++ src/services/mod.rs
@@ -1,4 +1,4 @@
 //! Services that are currently implemented.
 
-/// Reference: https://developer.valvesoftware.com/wiki/Master_Server_Query_Protocol#Format
+/// Reference: [Master Server Query Protocol](https://developer.valvesoftware.com/wiki/Master_Server_Query_Protocol)
 pub mod valve_master_server;