+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Games:
 - [Serious Sam](https://www.gog.com/game/serious_sam_the_first_encounter) support.
 - [Frontlines: Fuel of War](https://store.steampowered.com/app/9460/Frontlines_Fuel_of_War/) support.
 
+Services:
+- [Valve Master Server Query](https://developer.valvesoftware.com/wiki/Master_Server_Query_Protocol) support.
+
 ### Breaking:
 Protocols:
 - Valve: Request type enums have been renamed from all caps to starting-only uppercase, ex: `INFO` to `Info`